Add authorization checks when deleting/editing a post

+++ controllers/feed.js
@@ -145,6 +145,11 @@ exports.updatePost = (req, res, next) => {
             error.statusCode = 404;
             throw error;
         }
+        if(post.creator.toString() !== req.userId) {
+            const error = new Error('Not authorzied.');
+            error.statusCode = 403;
+            throw error;
+        }
         if(!imageUrl !== post.imageUrl) {
             clearImage(post.imageUrl);
         }
@@ -177,11 +182,16 @@ exports.deletePost = (req, res, next) => {
             throw error;
         }
         // checked logged in user
+        if(post.creator.toString() !== req.userId) {
+            const error = new Error('Not authorized.');
+            error.statusCode = 403;
+            throw error;
+        }
         clearImage(post.imageUrl);
         return Post.findByIdAndRemove(postId);      
     })
     .then(result => {
-        console.log(result);
+        //console.log(result);
         res.status(200).json({
             message: 'Deleted post.',           
         })
